+++ Sources/AppDelegate.swift
@@ -133,71 +133,6 @@ private enum CmuxThemeNotifications {
     static let reloadConfig = Notification.Name("com.cmuxterm.themes.reload-config")
 }
 
-enum GhosttyCrashBreadcrumb {
-    struct PendingCrash: Equatable {
-        let fileURL: URL
-        let modifiedAt: Date
-    }
-
-    static let lastCleanExitDefaultsKey = "ghosttyCrashBreadcrumb.lastCleanExitAt"
-    static let lastShownCrashDefaultsKey = "ghosttyCrashBreadcrumb.lastShownCrashAt"
-    static let notificationTabId = UUID(uuidString: "00000000-0000-0000-0000-000000003873")!
-
-    static var defaultCrashDirectoryURL: URL {
-        FileManager.default.homeDirectoryForCurrentUser
-            .appendingPathComponent(".local/state/ghostty/crash", isDirectory: true)
-    }
-
-    static func pendingCrash(
-        in crashDirectoryURL: URL = defaultCrashDirectoryURL,
-        defaults: UserDefaults = .standard,
-        fileManager: FileManager = .default
-    ) -> PendingCrash? {
-        guard let latest = latestCrashFile(in: crashDirectoryURL, fileManager: fileManager) else {
-            return nil
-        }
-
-        let lastCleanExit = defaults.object(forKey: lastCleanExitDefaultsKey) as? Date ?? .distantPast
-        let lastShownCrash = defaults.object(forKey: lastShownCrashDefaultsKey) as? Date ?? .distantPast
-        guard latest.modifiedAt > lastCleanExit, latest.modifiedAt > lastShownCrash else {
-            return nil
-        }
-        return latest
-    }
-
-    static func markShown(_ pendingCrash: PendingCrash, defaults: UserDefaults = .standard) {
-        defaults.set(pendingCrash.modifiedAt, forKey: lastShownCrashDefaultsKey)
-    }
-
-    static func markCleanExit(defaults: UserDefaults = .standard, date: Date = Date()) {
-        defaults.set(date, forKey: lastCleanExitDefaultsKey)
-    }
-
-    private static func latestCrashFile(
-        in crashDirectoryURL: URL,
-        fileManager: FileManager
-    ) -> PendingCrash? {
-        guard let urls = try? fileManager.contentsOfDirectory(
-            at: crashDirectoryURL,
-            includingPropertiesForKeys: [.contentModificationDateKey],
-            options: [.skipsHiddenFiles]
-        ) else {
-            return nil
-        }
-
-        return urls
-            .filter { $0.pathExtension == "ghosttycrash" }
-            .compactMap { url -> PendingCrash? in
-                guard let values = try? url.resourceValues(forKeys: [.contentModificationDateKey]),
-                      let modifiedAt = values.contentModificationDate else {
-                    return nil
-                }
-                return PendingCrash(fileURL: url, modifiedAt: modifiedAt)
-            }
-            .max { lhs, rhs in lhs.modifiedAt < rhs.modifiedAt }
-    }
-}
-
 func isCommandPaletteFocusStealingTerminalOrBrowserResponder(_ responder: NSResponder) -> Bool {
     if responder is GhosttyNSView || responder is WKWebView {
         return true
@@ -742,6 +677,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
     private var shortcutDefaultsObserver: NSObjectProtocol?
     private var menuBarVisibilityObserver: NSObjectProtocol?
     private var splitButtonTooltipRefreshScheduled = false
+    private var didScheduleGhosttyCrashBreadcrumbCheck = false
     private struct PendingConfiguredShortcutChord {
         let firstStroke: ShortcutStroke
         let windowNumber: Int?
@@ -1645,7 +1581,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
         self.tabManager = tabManager
         self.notificationStore = notificationStore
         self.sidebarState = sidebarState
-        showGhosttyCrashBreadcrumbIfNeeded(notificationStore: notificationStore)
+        scheduleGhosttyCrashBreadcrumbIfNeeded(notificationStore: notificationStore)
         disableSuddenTerminationIfNeeded()
         installLifecycleSnapshotObserversIfNeeded()
         prepareStartupSessionSnapshotIfNeeded()
@@ -1666,27 +1602,34 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
 #endif
     }
 
-    private func showGhosttyCrashBreadcrumbIfNeeded(notificationStore: TerminalNotificationStore) {
-        guard let pendingCrash = GhosttyCrashBreadcrumb.pendingCrash() else { return }
-        notificationStore.addNotification(
-            tabId: GhosttyCrashBreadcrumb.notificationTabId,
-            surfaceId: nil,
-            title: String(
-                localized: "crashBreadcrumb.title",
-                defaultValue: "cmux crashed during your last session"
-            ),
-            subtitle: String(
-                localized: "crashBreadcrumb.subtitle",
-                defaultValue: "Diagnostic file saved"
-            ),
-            body: String(
-                localized: "crashBreadcrumb.body",
-                defaultValue: "Diagnostic file saved to ~/.local/state/ghostty/crash/"
-            ),
-            cooldownKey: "ghosttyCrashBreadcrumb",
-            cooldownInterval: nil
-        )
-        GhosttyCrashBreadcrumb.markShown(pendingCrash)
+    private func scheduleGhosttyCrashBreadcrumbIfNeeded(notificationStore: TerminalNotificationStore) {
+        guard !didScheduleGhosttyCrashBreadcrumbCheck else { return }
+        didScheduleGhosttyCrashBreadcrumbCheck = true
+
+        Task { [weak notificationStore] in
+            guard let pendingCrash = await GhosttyCrashBreadcrumb.pendingCrashFromDefaultStorage(),
+                  !Task.isCancelled,
+                  let notificationStore else { return }
+            notificationStore.addNotification(
+                tabId: GhosttyCrashBreadcrumb.notificationTabId,
+                surfaceId: nil,
+                title: String(
+                    localized: "crashBreadcrumb.title",
+                    defaultValue: "cmux crashed during your last session"
+                ),
+                subtitle: String(
+                    localized: "crashBreadcrumb.subtitle",
+                    defaultValue: "Diagnostic file saved"
+                ),
+                body: String(
+                    localized: "crashBreadcrumb.body",
+                    defaultValue: "Diagnostic file saved to ~/.local/state/ghostty/crash/"
+                ),
+                cooldownKey: "ghosttyCrashBreadcrumb",
+                cooldownInterval: nil
+            )
+            GhosttyCrashBreadcrumb.markShown(pendingCrash)
+        }
     }
 
 #if DEBUG
